engine/tsr_model/requirements.txt
# TSR Model Dependencies
# 
# Note: These dependencies are now included in the main engine/requirements.txt
# Install from there with: pip install -r ../requirements.txt
#
# Key dependencies for this module:
# - torch (PyTorch neural networks)
# - pandas (data manipulation)
# - requests (API calls to Financial Modeling Prep) 
# - matplotlib, seaborn, plotly (visualizations)
# - python-dotenv (environment variables)
# - scikit-learn (ML utilities)

# For standalone installation of this module only:
torch
pandas
numpy
requests
python-dotenv
scikit-learn
matplotlib
seaborn
plotly
transformers
datasets